+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [5.0.0] 2024-06-17
+
 ### Added
 
 - Added ability to specify a custom max-evaluation-depth introduced with pySHACL 0.26.0